Reorganising and formatting
