+++ src/com/exolius/simplebackup/BackupHooks.java
@@ -2,14 +2,15 @@ package com.exolius.simplebackup;
 
 public class BackupHooks {
 
-    public void notifyBackupCreated(String command, String filename) {
-       if(!command.isEmpty()) {
-	  try {
-	    ProcessBuilder pb = new ProcessBuilder(command, filename);
-	    pb.start();
-	  } catch(Exception ex) {
-	    
-	  }
-       }
-    }
+	public void notifyBackupCreated(final String command, final String filename) {
+		if (!command.isEmpty()) {
+			try {
+				final ProcessBuilder pb = new ProcessBuilder(command, filename);
+				pb.start();
+			} catch (final Exception ex) {
+
+			}
+		}
+	}
+
 }
